requirements.txt
# Nexus-OS v1.1 依賴清單
# 核心 AI 框架
google-generativeai>=0.8.0

# Google API（日曆、Drive 等）
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# 資料庫（PostgreSQL + pgvector）
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.4

# 自然語言日期解析
dateparser>=1.2.0

# Web 框架
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# auth
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# 環境變數管理
python-dotenv>=1.0.0

# Telegram Bot（[webhooks] 為 Zeabur Webhook 模式必需）
python-telegram-bot[webhooks]>=20.7

# 排程師 Agent（APScheduler）
apscheduler>=3.10.0

# 天氣師 Agent（Open-Meteo）
openmeteo-requests>=1.2.0
requests-cache>=1.2.0
retry-requests>=2.0.0

# 工具函式庫
httpx>=0.27.0
pydantic>=2.0.0

# 爬虫師 Agent（BeautifulSoup + DuckDuckGo）
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0

# 圖片師 Agent（Imagen 4 + Gemini Vision 圖片處理）
Pillow>=10.0.0

# 金融分析師 Agent（Yahoo Finance 美股 + CoinGecko 加密貨幣）
yfinance>=0.2.40

# 加密貨幣交易 Agent（ccxt）
ccxt>=4.4.0

# Web3 Agent Wallet (Web3.py & Solana)
web3>=6.15.0
eth-account>=0.11.0
solana>=0.36.0
solders>=0.23.0
base58>=2.1.0
